pyproject.toml
[project]
name = "coin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.6.3"
dependencies = [
    "absl-py>=1.4.0",
    "aiohttp>=3.8.6",
    "certifi>=2025.4.26",
    "prometheus-client>=0.17.1",
    "websockets>=9.1",
]

[tool.uv.workspace]
members = [
    "cc/appcoin2/strategy/options_mm/script/wps",
]
